devices/sequencer/sequencer.configuration/pyproject.toml
[tool.poetry]
name = "sequencer-configuration"
version = "0.1.0"
description = ""
authors = ["damien <damien.bloch@institutoptique.fr>"]
readme = "README.md"
packages = [{include = "sequencer"}]

[tool.poetry.dependencies]
python = "^3.11"

device-config = {path = "../../device/device.configuration", develop = true}
settings-model = {path = "../../../common/settings_model", develop = true}
attrs = "^23.1.0"
serialization = {path = "../../../common/serialization", develop = true}


[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
